SelfOrganizingMaps

Documentation for SelfOrganizingMaps.

SelfOrganizingMaps.SquareSOM

SelfOrganizingMaps.euclideandistance

SelfOrganizingMaps.getBMUidx

SelfOrganizingMaps.getdiscpoints

SelfOrganizingMaps.getspherepoints

SelfOrganizingMaps.sphericaldistance

SelfOrganizingMaps.train!

SelfOrganizingMaps.updateWeights!

SelfOrganizingMaps.SquareSOM — Method

SquareSOM(nfeatures::Int, M::Int, N::Int, dist_func; η₀ = 0.1, # base λ = 0.1, σ₀² = 1.0, β = 0.1, nepochs = 25, wrap=:none)

Generate a Self Organizing Map with a square topology.

SelfOrganizingMaps.euclideandistance — Method

euclideandistance(i,j,m,n)

Compute the euclidean distance between grid points (i,j) and (m,n), i.e. d = (i-m)^2 + (j-n)^2

SelfOrganizingMaps.getBMUidx — Method

getBMUidx(som::SOM, x::AbstractVector)

Given a self organizing map som and input vector x, return the indices of the best matching unit (BMU).

SelfOrganizingMaps.getdiscpoints — Method

getdiscpoints(N::Int)

Generate a set of N (nearly) equally distributed points within a circle of radius 1.

SelfOrganizingMaps.getspherepoints — Method

getspherepoints(N::Int)

Generate N (nearly) evenly distributed points on the unit sphere using the fibonacci spiral procedure.

SelfOrganizingMaps.sphericaldistance — Method

sphericaldistance(ϕ₁,θ₁,ϕ₂,θ₂)

Compute the distance between two points on the unit sphere with coordinates (ϕ₁, θ₁), and (ϕ₂, θ₂) using the physics convention for ϕ the polar angle and θ the azimuth.

SelfOrganizingMaps.train! — Method

train!(som::SquareSOM, df::DataFrame)

Using the DataFrame df, train the weights of the som::SquareSOM.

SelfOrganizingMaps.updateWeights! — Method

updateWeights!(som::SquareSOM, x::AbstractVector)

Update the weights of the self organizing map som given the feature vector x.