Function Documentation

Documentation for SelfOrganizingMaps.

SelfOrganizingMaps.SelfOrganizingMap

SelfOrganizingMaps.SelfOrganizingMap — Type

SelfOrganizingMap

A model type for constructing a self organizing map, based on SelfOrganizingMaps.jl, and implementing the MLJ model interface.

From MLJ, the type can be imported using

SelfOrganizingMap = @load SelfOrganizingMap pkg=SelfOrganizingMaps

Do model = SelfOrganizingMap() to construct an instance with default hyper-parameters. Provide keyword arguments to override hyper-parameter defaults, as in SelfOrganizingMap(k=...).

SelfOrganizingMaps implements Kohonen's Self Organizing Map, Proceedings of the IEEE; Kohonen, T.; (1990):"The self-organizing map"

Training data

In MLJ or MLJBase, bind an instance model to data with mach = machine(model, X) where

X: an AbstractMatrix or Table of input features whose columns are of scitype Continuous.

Train the machine with fit!(mach, rows=...).

Hyper-parameters

k=10: Number of nodes along once side of SOM grid. There are k² total nodes.

η=0.5: Learning rate. Scales adjust made to wining node and its neighbors during each round of training.

σ²=0.05: The (squared) neighbor radius. Used to determine scale for neighbor node adjustments.

grid_type=:rectangular Node grid geometry. One of (:rectangular, :hexagonal, :spherical).

η_decay=:exponential Learning rate schedule function. One of (:exponential, :asymptotic)

σ_decay=:exponential Neighbor radius schedule function. One of (:exponential, :asymptotic, :none)

neighbor_function=:gaussian Kernel function used to make adjustment to neighbor weights. Scale is set by σ². One of (:gaussian, :mexican_hat).

matching_distance=euclidean Distance function from Distances.jl used to determine winning node.

Nepochs=1 Number of times to repeat training on the shuffled dataset.

Operations

transform(mach, X): returns the coordinates of the winning SOM node for each instance of X

Fitted parameters

The fields of fitted_params(mach) are:

weights: Array of weight vectors for the SOM nodes.

Report

coords: The coordinates of each of the SOM nodes.

Report

The fields of report(mach) are:

classes: the index of the winning node for each instance in X interpreted as a class label

Examples

using MLJ
som = @load SelfOrganizingMap pkg=SelfOrganizingMaps
model = som()
X, y = make_regression(50, 3) # synthetic data
mach = machine(model, X) |> fit!
X̃ = transform(mach, X)

rpt = report(mach)
classes = rpt.classes